+++ src/hooks/usePaymentSheet.tsx
@@ -10,6 +10,7 @@ export function usePaymentSheet() {
     initPaymentSheet: initPaymentSheetNative,
     presentPaymentSheet: presentPaymentSheetNative,
     confirmPaymentSheetPayment: confirmPaymentSheetPaymentNative,
+    resetPaymentSheetCustomer: resetPaymentSheetCustomerNative,
   } = useStripe();
   const [loading, setLoading] = useState(false);
 
@@ -34,10 +35,23 @@ export function usePaymentSheet() {
     return result;
   };
 
+  const resetPaymentSheetCustomer = async () => {
+    setLoading(true);
+    const result = await resetPaymentSheetCustomerNative();
+    setLoading(false);
+    return result;
+  };
+
   return {
     loading,
     initPaymentSheet,
     presentPaymentSheet,
     confirmPaymentSheetPayment,
+    /**
+     * You must call this method when the user logs out from your app. This will ensure that
+     * any persisted authentication state in the PaymentSheet, such as authentication cookies,
+     * is also cleared during logout.
+     */
+    resetPaymentSheetCustomer,
   };
 }